fix data type
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@114 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Frameworks/LuceneKit/Headers/Index/LCTermEnum.h
@@ -11,7 +11,7 @@
 
 - (BOOL) next;
 - (LCTerm *) term;
-- (int) docFreq;
+- (long) docFreq;
 - (void) close;
 - (BOOL) skipTo: (LCTerm *) target;
